Move countActiveSessions to UserSessionRepository
